Revert "LG-3840 Updated welcome and new agreement pages (#4811)"
This reverts commit 24dcf629e7e3d29c78b9721e2936768bfc1392e1.
