+++ config/locales/doc_auth/en.yml
@@ -90,10 +90,8 @@ en:
       document_capture_selfie: Your photo
       front: Front
       interstitial: We are processing your images
-      lets_go: Let's get started
       photo: Photo
       review_issues: Check your images and try again
-      secure_account: Secure your account
       selfie: Take a photo of yourself
       ssn: Please enter your social security number.
       take_picture: Take a photo with a phone
@@ -104,8 +102,7 @@ en:
         a photo of yourself
       upload_liveness_enabled: How would you like to verify your identity?
       verify: Please verify your information
-      verify_identity: Verify your identity
-      welcome: Verify your identity to securely access government services
+      welcome: We need to verify your identity
     info:
       camera_required: Your mobile phone must have a camera and a web browser
       capture_status_capturing: Capturing
@@ -125,7 +122,6 @@ en:
         when it’s done.
       interstitial_thanks: Thanks for your patience!
       keep_window_open: Do not close this window.
-      lets_go: 'Identity verification happens in two parts:'
       link_sent: Please check your phone and follow instructions to take a photo of
         your state issued ID.
       link_sent_complete_js: The next step will load automatically once you verify
@@ -134,12 +130,6 @@ en:
         your identity.
       no_other_id_help_bold_html: "<strong>If you do not have a state-issued ID</strong>,
         <a href=%{failure_to_proof_url}>get help at %{sp_name}.</a>"
-      privacy_html: Login.gov is a secure, government website that adheres to the
-        highest standards in data protection. We only use your data to verify your
-        identity. %{link} about our privacy and security measures.
-      secure_account: After you verify, we'll ask you to encrypt your account. Encryption
-        means your data is protected and only you, the account holder, will be able
-        to access or change your information.
       tag: Recommended
       take_picture: Use the camera on your mobile phone and upload images of your
         ID.  We only use the images to verify your identity.
@@ -156,17 +146,16 @@ en:
         identity.
       use_cac: Do you have a government employee ID?
       use_cac_link: Use a PIV/CAC instead
-      verify_identity: We'll ask for your personal information. We'll use, keep and
-        share some of your personal information to verify your identity against public
-        records.
       verifying: Verifying…
-      welcome_html: The agency that you are trying to access needs to make sure you
-        are you &#8212; not someone pretending to be you.
+      welcome: We verify your identity to make sure you are you—not someone pretending
+        to be you.  Verifying your identity lets you access services that handle sensitive
+        information.
     instructions:
-      bullet1: Your State-issued ID
-      bullet1a: A device with a camera
-      bullet2: Your Social Security number
-      bullet3: A phone number on a phone plan associated with your name
+      bullet1: Upload your ID
+      bullet1a: Take a photo of yourself
+      bullet2: Confirm your address
+      bullet3: Secure your account
+      bullet4: Account recovery
       consent: By checking this box, you are letting login.gov ask for, use, keep,
         and share your personal information. We will only use it to verify your identity.
       document_capture_selfie_consent_action: "<lg-underline>Allow access to the camera</lg-underline>
@@ -184,17 +173,15 @@ en:
       email_sent: Link sent to %{email}. Please check your desktop email and follow
         instructions to verify your identity.
       learn_more: Learn more
-      privacy: Our privacy and security standards
       send_sms: We'll send a text message to your device with a link.  Follow that
         link to your browser to take photos of the front and back of your ID.
       switch_back: Switch back to your computer to finish verifying your identity
-      text1: ''
-      text1a: such as a phone or computer.
-      text2: You will not need the card with you.
-      text3: You do not need to be the primary account holder. If you do not have
-        a phone number that qualifies, we can send you the verification code by mail
-        in approximately 3-7 business days.
-      welcome: 'To verify your identity, you will need:'
+      text1: along with your Social Security Number
+      text1a: with your phone or webcam
+      text2: using a phone number or mailing address
+      text3: by re-entering your login.gov password
+      text4: make sure you always have access
+      welcome: 'What you''ll need to do:'
     tips:
       document_capture_header_text: 'For best results:'
       document_capture_hint: Must be a JPG, BMP, PNG, or TIFF
