+++ config/locales/doc_auth/es.yml
@@ -97,10 +97,8 @@ es:
       document_capture_selfie: Su foto
       front: Parte Delantera
       interstitial: Estamos procesando sus imágenes
-      lets_go: Comencemos
       photo: Foto
       review_issues: Revise sus imágenes e inténtelo de nuevo
-      secure_account: Asegure su cuenta
       selfie: Tómese una foto
       ssn: Por favor ingrese su número de seguro social.
       take_picture: Toma una foto con un teléfono
@@ -111,9 +109,7 @@ es:
         de identidad y tomarse
       upload_liveness_enabled: "¿Cómo te gustaría verificar su identidad?"
       verify: Por favor verifica tu información
-      verify_identity: Verifique su identidad
-      welcome: Verifique su identidad para acceder de forma segura a los servicios
-        gubernamentales
+      welcome: Nosotros necesitamos verificar tu identidad
     info:
       camera_required: Su teléfono móvil debe tener una cámara y un navegador web
       capture_status_capturing: Capturando
@@ -134,7 +130,6 @@ es:
         paso automáticamente cuando esté terminado.
       interstitial_thanks: "¡Gracias por su paciencia!"
       keep_window_open: No cierres esta ventana.
-      lets_go: 'La verificación de la identidad se realiza en dos partes:'
       link_sent: Verifique su teléfono y siga las instrucciones para tomar una fotografía
         de la identificación emitida por su estado.
       link_sent_complete_js: El siguiente paso se cargará automáticamente una vez
@@ -143,13 +138,6 @@ es:
         la verificación de tu identidad.
       no_other_id_help_bold_html: "<strong>Si no tiene un documento de identidad expedido
         por el estado</strong>, <a href=%{failure_to_proof_url}>obtenga ayuda en %{sp_name}.</a>"
-      privacy_html: Login.gov es un sitio web gubernamental seguro que cumple con
-        las normas más estrictas de protección de datos. Solo utilizamos sus datos
-        para verificar su identidad. %{link} sobre nuestras medidas de privacidad
-        y seguridad.
-      secure_account: Una vez que haya verificado su identidad, le pediremos que cifre
-        su cuenta. Con el cifrado, sus datos quedan protegidos, por lo que solo usted,
-        el titular de la cuenta, puede acceder a su información o modificarla.
       tag: Recomendado
       take_picture: Use la cámara en su teléfono móvil y cargue imágenes de su identificación.
         Solo usamos las imágenes para verificar su identidad.
@@ -168,19 +156,18 @@ es:
         su identidad.
       use_cac: "¿Tiene un documento de identidad de empleado del gobierno?"
       use_cac_link: Utilice un PIV/CAC en su lugar
-      verify_identity: Le pediremos su información personal. Utilizaremos, conservaremos
-        y compartiremos algunos de sus datos personales para verificar su identidad
-        con los registros públicos.
       verifying: Verificando…
-      welcome_html: La agencia a la que está intentando acceder debe asegurarse de
-        que usted sea quien dice ser, y no alguien que se hace pasar por usted.
+      welcome: Verificamos su identidad para asegurarnos de que usted es usted, y
+        no alguien que pretende ser usted. Verificar su identidad le permite acceder
+        a servicios que manejan información confidencial.
     instructions:
-      bullet1: Su documento de identidad emitido por el estado.
-      bullet1a: Un dispositivo con cámara
-      bullet2: Su Número de Seguridad Social.
-      bullet3: Un número de teléfono con un plan tarifario vinculado a su nombre.
-      consent: Al marcar esta casilla, usted permite que login.gov solicite, utilice,
-        conserve y comparta su información personal. Solo la utilizaremos para verificar
+      bullet1: Sube tu ID
+      bullet1a: Tómese una foto
+      bullet2: Confirma tu direccion
+      bullet3: Asegure su cuenta
+      bullet4: Recuperación de cuenta
+      consent: Al marcar esta casilla, está permitiendo que login.gov solicite, use,
+        conserve y comparta su información personal. Solo lo usaremos para verificar
         su identidad.
       document_capture_selfie_consent_action: "<lg-underline>Permita el acceso a la
         cámara</lg-underline> y busque una notificación en su navegador o revise la
@@ -196,19 +183,17 @@ es:
         con la imagen en el frente de su identificación.
       email_sent: Enlace enviado a %{email}. Compruebe el correo electrónico de su
         escritorio y siga las instrucciones para verificar su identidad.
-      learn_more: Obtenga más información
-      privacy: Nuestras normas de privacidad y seguridad
+      learn_more: Aprende más
       send_sms: Le enviaremos un mensaje de texto a su dispositivo con un enlace.
         Siga ese enlace a su navegador para tomar fotografías de la parte delantera
         y trasera de su identificación.
       switch_back: Vuelve a tu computadora para terminar de verificar tu identidad
-      text1: ''
-      text1a: como un teléfono o una computadora.
-      text2: No es necesario disponer de la credencial.
-      text3: No es necesario que sea el titular principal de la cuenta. Si no tiene
-        un número de teléfono que cumpla con este requisito, podemos enviarle el código
-        de verificación por correo en un plazo aproximado de 3 a 7 días laborables.
-      welcome: 'Deberá contar con lo siguiente para verificar su identidad:'
+      text1: junto con su número de seguro social
+      text1a: con un teléfono o cámara web
+      text2: usando un número de teléfono o dirección de correo
+      text3: volviendo a ingresar su contraseña de login.gov
+      text4: asegúrate de tener siempre acceso
+      welcome: 'Lo que necesitarás hacer:'
     tips:
       document_capture_header_text: 'Para obtener los mejores resultados:'
       document_capture_hint: Debe ser un JPG, BMP, PNG o TIFF
