+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_03_16_082419) do
+ActiveRecord::Schema.define(version: 2021_03_15_144559) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -189,8 +189,6 @@ ActiveRecord::Schema.define(version: 2021_03_16_082419) do
     t.integer "document_capture_view_count", default: 0
     t.integer "document_capture_submit_count", default: 0
     t.integer "document_capture_error_count", default: 0
-    t.datetime "agreement_view_at"
-    t.integer "agreement_view_count", default: 0
     t.index ["user_id"], name: "index_doc_auth_logs_on_user_id", unique: true
     t.index ["verified_view_at"], name: "index_doc_auth_logs_on_verified_view_at"
   end
